+++ .gitignore
@@ -1,5 +1,5 @@
 dist
-site
+public
 coverage
 node_modules
 package-lock.json
