+++ packages/components/notes.md
@@ -1,0 +1,34 @@
+# notes
+
+mdx-deck v2 prototype
+
+todo:
+
+- [x] Head
+- [x] Image
+- [x] Notes
+- [x] Appear
+- [ ] Code (may handle this with theming)
+- [ ] normalize slide styles with v1
+- [x] history api fallback
+- [x] mdx components
+- [x] themes
+- [ ] layouts
+- [ ] presenter mode
+  - [ ] timer/clock
+  - [ ] open new tab button
+  - [ ] look at google slides behavior
+- [ ] overview mode
+  - [ ] highlighted state bug
+  - [ ] slide numbers
+- [ ] localStorage
+- [ ] keyboard shortcuts
+- [ ] docs for customizing any component
+- [x] swipeable
+
+extras
+
+- [ ] docs for syntax highlighting
+- [ ] Print view
+- [ ] PDF export?
+- [ ] dots??
