+++ packages/components/src/Overview.js
@@ -1,92 +1,65 @@
 import React from 'react'
-import PropTypes from 'prop-types'
-import get from 'lodash.get'
-import Box from './Box'
-import Flex from './Flex'
+import { Link } from '@reach/router'
 import Zoom from './Zoom'
 import Slide from './Slide'
-import Root from './Root'
-import Mono from './Mono'
 
-export const Overview = ({
-  index,
-  length,
-  slides = [],
-  mode,
-  metadata = {},
-  update,
-  step,
-  ...props
-}) => {
-  const notes = get(metadata, index + '.notes')
+const noop = () => {}
+
+export const Overview = props => {
+  const { index, slides } = props
 
   return (
-    <Flex
-      color="white"
-      bg="black"
-      css={{
+    <div
+      style={{
+        display: 'flex',
         alignItems: 'flex-start',
         height: '100vh',
+        backgroundColor: 'black',
       }}
     >
-      <Box
-        mr="auto"
-        px={2}
-        py={3}
-        css={{
+      <div
+        style={{
           flex: 'none',
           height: '100vh',
+          paddingLeft: 4,
+          paddingRight: 4,
           overflowY: 'auto',
+          marginRight: 'auto',
         }}
       >
         {slides.map((Component, i) => (
-          <Box
+          <Link
             key={i}
-            role="link"
-            p={1}
+            to={'/' + i}
             style={{
-              outline: i === index ? '1px solid #07c' : null,
-            }}
-            css={{
+              display: 'block',
+              color: 'inherit',
+              textDecoration: 'none',
+              padding: 0,
+              marginTop: 4,
+              marginBottom: 4,
               cursor: 'pointer',
-            }}
-            onClick={e => {
-              update({ index: i })
+              outline: i === index ? '4px solid #0cf' : null,
             }}
           >
             <Zoom zoom={1 / 6}>
-              <Root {...props}>
-                <Slide>
-                  <Component />
-                </Slide>
-              </Root>
+              <Slide register={noop}>
+                <Component />
+              </Slide>
             </Zoom>
-          </Box>
+          </Link>
         ))}
-      </Box>
-      <Box mx="auto" py={4} width={2 / 3}>
-        <Zoom zoom={2 / 3}>
-          <Root {...props}>{props.children}</Root>
-        </Zoom>
-        <Flex>
-          <Box ml="auto" py={2}>
-            {index + 1}/{length}
-          </Box>
-        </Flex>
-        <Box mt={3}>{notes}</Box>
-      </Box>
-    </Flex>
+      </div>
+      <div
+        style={{
+          width: 200 / 3 + '%',
+          margin: 'auto',
+        }}
+      >
+        <Zoom zoom={2 / 3}>{props.children}</Zoom>
+      </div>
+    </div>
   )
-}
-
-Overview.propTypes = {
-  index: PropTypes.number.isRequired,
-  length: PropTypes.number.isRequired,
-  update: PropTypes.func.isRequired,
-  step: PropTypes.number.isRequired,
-  slides: PropTypes.array,
-  mode: PropTypes.string,
-  notes: PropTypes.object,
 }
 
 export default Overview
