+++ packages/components/src/_ref/SlideDeck.js
