+++ packages/components/src/_temp.js
@@ -1,0 +1,8 @@
+// may cut some of this
+
+const GoogleFonts = withTheme(
+  props =>
+    !!props.theme.googleFont && (
+      <link href={props.theme.googleFont} rel="stylesheet" />
+    )
+)
