+++ packages/html-plugin/README.md
@@ -1,0 +1,16 @@
+# @mdx-deck/webpack-html-plugin
+
+Webpack plugin for generating HTML
+
+```sh
+npm i @mdx-deck/webpack-html-plugin
+```
+
+```js
+// webpack.config.js
+const HTMLPlugin = require('@mdx-deck/webpack-html-plugin')
+
+module.exports = {
+  plugins: [new HTMLPlugin()],
+}
+```
