+++ packages/html-plugin/package.json
@@ -3,5 +3,8 @@
   "version": "2.0.0-0",
   "author": "Brent Jackson <jxnblk@gmail.com>",
   "license": "MIT",
-  "repository": "github:jxnblk/mdx-deck"
+  "repository": "github:jxnblk/mdx-deck",
+  "dependencies": {
+    "webpack-sources": "^1.3.0"
+  }
 }
