+++ packages/layouts/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@mdx-deck/components",
+  "name": "@mdx-deck/layouts",
   "version": "2.0.0-0",
-  "main": "src/index.js",
+  "main": "index.js",
   "author": "Brent Jackson <jxnblk@gmail.com>",
   "license": "MIT"
 }
