+++ packages/loader/index.js
@@ -1,22 +1,18 @@
 const { getOptions } = require('loader-utils')
 const mdx = require('@mdx-js/mdx')
-// const normalizeNewline = require('normalize-newline')
-const mdxPluginSplit = require('mdx-plugin-split')
+const mdxPlugin = require('@mdx-deck/mdx-plugin')
 
 module.exports = async function(src) {
   const callback = this.async()
   const options = getOptions(this) || {}
-  // options.skipExport = true
   options.mdPlugins = options.mdPlugins || []
-  options.mdPlugins.push(mdxPluginSplit)
+  options.mdPlugins.push(mdxPlugin)
 
   const result = mdx.sync(src, options)
 
   const code = `import React from 'react'
   import { MDXTag } from '@mdx-js/tag'
   ${result}`
-
-  console.log(code)
 
   return callback(null, code)
 }
