+++ packages/loader/package.json
@@ -1,0 +1,12 @@
+{
+  "name": "@mdx-deck/loader",
+  "version": "2.0.0-0",
+  "main": "index.js",
+  "author": "Brent Jackson <jxnblk@gmail.com>",
+  "license": "MIT",
+  "dependencies": {
+    "@mdx-deck/mdx-plugin": "^2.0.0-0",
+    "@mdx-js/mdx": "^0.20.0",
+    "loader-utils": "^1.2.3"
+  }
+}
