+++ packages/themes/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@mdx-deck/components",
+  "name": "@mdx-deck/themes",
   "version": "2.0.0-0",
   "main": "src/index.js",
   "author": "Brent Jackson <jxnblk@gmail.com>",
